Correction advancements
This is a trial at correcting the app that didn't work.

When we run the inference, the output data gives us floatArrays of 425 elements and I can't figure out how our 80 classes transformed into 425 data.
I hypothesized it was 85 * 5, which is (80 nbclasses + 4 box coordinates + 1 confidence) but I'm not sure about anything and can't get it right.